requirements.txt
# API and Web Framework
fastapi==0.100.0
uvicorn[standard]==0.23.2
pydantic==2.1.1
starlette==0.27.0
python-multipart==0.0.6
httpx==0.24.1

# Database and Storage
sqlalchemy==2.0.19
aiosqlite==0.19.0
alembic==1.11.2
redis==4.6.0
pyyaml==6.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Async Utilities
asyncio==3.4.3
aiofiles==23.1.0
aiohttp==3.8.5
websockets==11.0.3

# Deployment and Infrastructure
docker==6.1.3
kubernetes==27.2.0
prometheus-client==0.17.1

# Development and Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
mock==5.1.0

# Utilities
python-dateutil==2.8.2
tenacity==8.2.2
schedule==1.2.0
structlog==23.1.0
